Swap meanings of `assert_not_running` and `assert_running`.
In the old definition, `assert_running` would fail if the server was running,
and `assert_not_running` would fail if the server was not running.

In other words, `assert_running` == "raise an assertion if we are running". I think this could be more clear if we wrote `assert_if_running` instead.

My background is from languages like Python,
where I'd expect `assert_running` would assert the server is running, and fail otherwise.

To make the script more legible to me,
I'm inverting the meanings of the two helpers.