Create .gitignore
It would be convenient if we could set up this script by cloning this repository, and then immediately running the script.

However, this will download a bunch of server files into the directory as well, and we'd like to be able to keep using the directory as a git repository without extra effort, so we set up a gitignore to automatically exclude everything but the files originally included in the repo.