Add QoL improvements, fix minor bugs
This commit adds some quality of life improvements like a warning in the server chat before the server shuts down, and the ability to overwrite the service symlinks if they already exist.

It also reorders the sequence of steps in the `init` phase so that backups are initialized before the backup service starts, and so that the "`source .bashrc`" message prints last.

There was a minor syntax error in the handling of the "Run source bashrc" message that would appear when the BASHRC_MODIFIED variable was not set; this commit quotes the variable correctly to avoid this.