Allow services to run with configurable user and group
By default, the services run with the `minecraft` user and group, but there's no reason we can't let people customize this to be any user they like.

This commit also repairs a bug in which the working directory of the services wasn't set properly, causing the service to fail to start.